[SP] Fix spawn field / vehicle field offset not being size_t storage type.

+++ codeJK2/game/g_spawn.cpp
@@ -179,7 +179,7 @@ typedef enum {
 typedef struct
 {
 	char	*name;
-	int		ofs;
+	size_t		ofs;
 	fieldtype_t	type;
 	int		flags;
 } field_t;
